requirements.txt
requests>=2.32
pandas>=2.3
numpy>=2.4
matplotlib>=3.10
seaborn>=0.13
scikit-learn>=1.8
pyyaml>=6.0
pytest>=8.0
